reliable: Workaround missing context.Context and use custom readTimeout
Now there is a big difference between the entire operation should not
run more than 2s and each I/O operation might not take more than 400ms.
It is not clear how many I/O ops are needed to make a proper health
check, not all of them are equal (e.g. heavy query vs. authentication)
but it seems the only way to get it going.